+++ src/olmo_core/data/data_loader.py
@@ -35,9 +35,11 @@ import numpy as np
 import torch
 import torch.distributed as dist
 import torch.utils.data
+from torch.distributed import DeviceMesh
 
 from ..aliases import PathOrStr
 from ..config import Config
+from ..distributed.parallel import get_dp_process_group
 from ..distributed.utils import barrier, get_fs_local_rank, get_rank, get_world_size
 from ..exceptions import OLMoConfigurationError
 from ..utils import get_default_device, roundrobin, threaded_generator
@@ -430,9 +432,14 @@ class NumpyDataLoaderBase(DataLoaderBase):
         self.build_and_save_global_indices(in_memory=in_memory)
 
     def get_mock_batch(self) -> Dict[str, Any]:
+        rng = torch.Generator()
+        rng.manual_seed(self.seed + self.dp_rank)
         num_instances = self.rank_batch_size // self.dataset.max_sequence_length
         input_ids = torch.randint(
-            0, self.dataset.vocab_size, (num_instances, self.dataset.max_sequence_length)
+            0,
+            self.dataset.vocab_size,
+            (num_instances, self.dataset.max_sequence_length),
+            generator=rng,
         )
         return {"input_ids": input_ids}
 
@@ -908,13 +915,24 @@ class NumpyDataLoaderConfig(Config):
         dataset: NumpyDatasetBase,
         *,
         collator: Optional[DataCollator] = None,
+        mesh: Optional[DeviceMesh] = None,
         dp_process_group: Optional[dist.ProcessGroup] = None,
     ) -> NumpyDataLoaderBase:
         """
         Construct the :class:`NumpyDataLoaderBase`.
+
+        :param dataset: The dataset.
+        :param mesh: An optional ``DeviceMesh`` that defines the data parallel dimensions. Ideally
+            you should create this mesh using :func:`~olmo_core.distributed.parallel.build_device_mesh()`
+            or equivalently :meth:`olmo_core.nn.transformer.TransformerConfig.build_mesh()`.
+            Alternatively you can pass the ``dp_process_group`` instead.
+        :param dp_process_group: The data parallel process group.
         """
         if self.work_dir is not None and not dataset.work_dir_set:
             dataset.work_dir = Path(self.work_dir)
+
+        if dp_process_group is None and mesh is not None:
+            dp_process_group = get_dp_process_group(mesh)
 
         dataset.prepare()