+++ src/olmo_core/internal/model_ladder.py
@@ -7,6 +7,7 @@ from rich import print
 
 from olmo_core.config import Config, StrEnum
 from olmo_core.data import NumpyDataLoaderConfig, NumpyDatasetConfig
+from olmo_core.distributed.utils import get_local_rank
 from olmo_core.launch.beaker import BeakerLaunchConfig
 from olmo_core.model_ladder import ModelLadder, ModelSize
 from olmo_core.nn.transformer import TransformerConfig
@@ -48,8 +49,9 @@ class SubCmd(StrEnum):
         else:
             raise NotImplementedError(self)
 
-    def run(self, size: ModelSize, config: LadderRunConfig):
-        print(config)
+    def run(self, config: LadderRunConfig):
+        if get_local_rank() == 0:
+            print(config)
 
         if self == SubCmd.launch:
             config.launch.launch(follow=True)
@@ -60,17 +62,22 @@ class SubCmd(StrEnum):
                 # Set RNG states on all devices.
                 seed_all(config.ladder.init_seed)
 
+                device = get_default_device()
+
+                # Build mesh, if needed.
+                world_mesh = config.model.build_mesh(device=device)
+
                 # Build components.
                 model = config.model.build(
                     init_device="meta",
-                    device=get_default_device(),
+                    device=device,
                     max_seq_len=config.dataset.sequence_length,
-                    dp_mesh=config.ladder.get_dp_mesh(size=size),
+                    mesh=world_mesh,
                 )
                 optim = config.optim.build(model)
                 dataset = config.dataset.build()
-                data_loader = config.data_loader.build(dataset)
-                trainer = config.trainer.build(model, optim, data_loader)
+                data_loader = config.data_loader.build(dataset, mesh=world_mesh)
+                trainer = config.trainer.build(model, optim, data_loader, mesh=world_mesh)
 
                 # Record the config to W&B/Comet and each checkpoint dir.
                 config_dict = config.as_config_dict()
@@ -163,4 +170,4 @@ $ [i]python {sys.argv[0]} {SubCmd.launch} 1B ai2/pluto-cirrascale --launch.num_n
     config.ladder.validate()
 
     # Run the cmd.
-    cmd.run(size, config)
+    cmd.run(config)