+++ src/olmo_core/nn/feed_forward.py
@@ -5,11 +5,15 @@ from typing import Optional
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+from torch.distributed import DeviceMesh
+from torch.distributed.tensor.parallel import parallelize_module
+from torch.distributed.tensor.placement_types import Placement
 
 from ..config import Config, DType, StrEnum
 from ..doc_utils import beta_feature
 from ..exceptions import OLMoConfigurationError
 from .functional import l2_normalize
+from .utils import get_tp_wrappers
 
 __all__ = ["FeedForwardType", "FeedForwardConfig", "FeedForward", "NormalizedFeedForward"]
 
@@ -117,6 +121,27 @@ class FeedForward(nn.Module):
         """
         return self.w2(F.silu(self.w1(x)) * self.w3(x))
 
+    def apply_tp(
+        self,
+        tp_mesh: DeviceMesh,
+        output_layouts: Optional[Placement] = None,
+        use_local_output: bool = True,
+        float8_enabled: bool = False,
+    ):
+        rowwise_parallel, colwise_parallel, _ = get_tp_wrappers(float8_enabled=float8_enabled)
+
+        parallelize_module(
+            module=self,
+            device_mesh=tp_mesh,
+            parallelize_plan={
+                "w1": colwise_parallel(),
+                "w2": rowwise_parallel(
+                    output_layouts=output_layouts, use_local_output=use_local_output
+                ),
+                "w3": colwise_parallel(),
+            },
+        )
+
 
 @beta_feature
 class NormalizedFeedForward(FeedForward):
@@ -159,6 +184,19 @@ class NormalizedFeedForward(FeedForward):
         sw1 = self.sw1 * ((self.sw_init_value / self.sw_init_scaling) * self.sqrt_d_model)
         sw3 = self.sw3 * (self.sw_init_value / self.sw_init_scaling)
         return self.w2(F.silu(sw1 * self.w1(x)) * (sw3 * self.w3(x)))
+
+    def apply_tp(
+        self,
+        tp_mesh: DeviceMesh,
+        output_layouts: Optional[Placement] = None,
+        use_local_output: bool = True,
+        float8_enabled: bool = False,
+    ):
+        del tp_mesh, output_layouts, use_local_output, float8_enabled
+
+        raise NotImplementedError(
+            "TP is not implemented yet for the normalized feed-forward variant"
+        )
 
     @torch.no_grad()
     def normalize_matrices(self):