+++ src/scripts/train/OLMo2-26B.py
@@ -1,30 +1,47 @@
 """
-Train a 1B OLMo model. Run this script without any arguments to see usage info.
+Train a 26B OLMo model. Run this script without any arguments to see usage info.
 """
+
+import logging
 
 from olmo_core.config import DType
 from olmo_core.distributed.parallel import DataParallelType
+from olmo_core.float8 import Float8Config
 from olmo_core.internal.experiment import CommonComponents, main
-from olmo_core.nn.transformer import TransformerConfig, TransformerDataParallelConfig
+from olmo_core.nn.transformer import (
+    TransformerActivationCheckpointingConfig,
+    TransformerActivationCheckpointingMode,
+    TransformerConfig,
+    TransformerDataParallelConfig,
+)
 from olmo_core.optim import AdamWConfig, OptimGroupOverride
 from olmo_core.train import TrainerConfig
 from olmo_core.train.callbacks import CheckpointerCallback, CometCallback, WandBCallback
 
+log = logging.getLogger(__name__)
+
 
 def build_model_config(common: CommonComponents) -> TransformerConfig:
-    return TransformerConfig.olmo2_1B(
+    compile = True
+    return TransformerConfig.olmo2_26B(
         vocab_size=common.tokenizer.padded_vocab_size(),
-        compile=True,
+        compile=compile,
+        fused_ops=False,
+        use_flash=not compile,
         dp_config=TransformerDataParallelConfig(
-            name=DataParallelType.hsdp, param_dtype=DType.bfloat16, reduce_dtype=DType.float32
+            name=DataParallelType.fsdp, param_dtype=DType.bfloat16, reduce_dtype=DType.float32
         ),
+        ac_config=TransformerActivationCheckpointingConfig(
+            mode=TransformerActivationCheckpointingMode.full
+        ),
+        float8_config=Float8Config(compile=compile, enabled=False),
     )
 
 
 def build_optim_config(common: CommonComponents) -> AdamWConfig:
     del common
     return AdamWConfig(
-        lr=4e-4,
+        lr=6e-4,
         weight_decay=0.1,
         betas=(0.9, 0.95),
         group_overrides=[
@@ -38,7 +55,7 @@ def build_trainer_config(common: CommonComponents) -> TrainerConfig:
     return (
         TrainerConfig(
             save_folder=common.save_folder,
-            rank_microbatch_size=8 * 4096,
+            rank_microbatch_size=4 * 4096,
             save_overwrite=True,
             metrics_collect_interval=10,
             cancel_check_interval=1,
@@ -49,7 +66,7 @@ def build_trainer_config(common: CommonComponents) -> TrainerConfig:
             "checkpointer",
             CheckpointerCallback(
                 save_interval=10_000,
-                ephemeral_save_interval=1000,
+                ephemeral_save_interval=250,
                 save_async=True,
             ),
         )
@@ -58,7 +75,7 @@ def build_trainer_config(common: CommonComponents) -> TrainerConfig:
             CometCallback(
                 name=common.run_name,
                 workspace="ai2",
-                project="OLMo-core-1B",
+                project="OLMo-core-26B",
                 enabled=True,
                 cancel_check_interval=10,
             ),
@@ -68,7 +85,7 @@ def build_trainer_config(common: CommonComponents) -> TrainerConfig:
             WandBCallback(
                 name=common.run_name,
                 entity="ai2-llm",
-                project="OLMo-core-1B",
+                project="OLMo-core-26B",
                 enabled=False,
                 cancel_check_interval=10,
             ),
@@ -78,7 +95,7 @@ def build_trainer_config(common: CommonComponents) -> TrainerConfig:
 
 if __name__ == "__main__":
     main(
-        global_batch_size=1024 * 4096,
+        global_batch_size=2048 * 4096,
         model_config_builder=build_model_config,
         optim_config_builder=build_optim_config,
         trainer_config_builder=build_trainer_config,